Remove MemoryField and CPUField definitions from models.py and add them to the new fields.py
